+++ server/collaboration/PersistenceExtension.ts
@@ -9,8 +9,8 @@ import { sequelize } from "@server/database/sequelize";
 import Logger from "@server/logging/Logger";
 import { trace } from "@server/logging/tracing";
 import Document from "@server/models/Document";
+import ProsemirrorHelper from "@server/models/helpers/ProsemirrorHelper";
 import documentCollaborativeUpdater from "../commands/documentCollaborativeUpdater";
-import markdownToYDoc from "./utils/markdownToYDoc";
 
 @trace()
 export default class PersistenceExtension implements Extension {
@@ -51,11 +51,11 @@ export default class PersistenceExtension implements Extension {
         "database",
         `Document ${documentId} is not in state, creating from markdown`
       );
-      const ydoc = markdownToYDoc(document.text, fieldName);
-      const state = Y.encodeStateAsUpdate(ydoc);
+      const ydoc = ProsemirrorHelper.toYDoc(document.text, fieldName);
+      const state = ProsemirrorHelper.toState(ydoc);
       await document.update(
         {
-          state: Buffer.from(state),
+          state,
         },
         {
           silent: true,
